requirements.txt
cbor2
cryptography
pympler
python-dateutil

beautifulsoup4>=4.0.0
chardet>=3.0.4
cloudscraper>=1.2.33
html5lib>=1.0.1
lxml>=4.5.0
lz4>=3.0.0
nose>=1.3.0
numpy>=1.16.0
Pillow>=6.0.0
psutil>=5.0.0
pyOpenSSL>=19.1.0
PySocks>=1.7.0
PyYAML>=5.0.0
Send2Trash>=1.5.0
service-identity>=18.1.0
six>=1.14.0
Twisted>=20.3.0

opencv-python-headless==4.5.5.64
python-mpv==1.0.3
requests==2.31.0

QtPy==2.3.0
PySide6==6.4.1

setuptools==65.5.1

gimpformats